server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Jonathan van Wersch <jonathanvwersch@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pydantic = "^2.0.0"
sqlalchemy = ">=1.4,<2.0.36"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
supabase = "^2.9.1"
pyjwt = "^2.9.0"
python-multipart = "^0.0.17"
langchain = "^0.3.7"
langchain-openai = "^0.2.8"
pypdf = "^5.1.0"
python-magic = "^0.4.27"
langchain-community = "^0.3.7"
openai = "^1.0.0"
marker-pdf = "^0.3.10"
voyageai = "^0.3.1"
fitz = "^0.0.1.dev2"
pymupdf = "^1.24.14"
mypy = "^1.13.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
